requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
supabase>=2.0.0
psycopg2-binary>=2.9.0
mcp>=0.9.0
fastapi>=0.104.0
uvicorn>=0.24.0
sse-starlette>=1.8.0

perplexityai
